+++ src/main/java/life/majiang/community/controller/QuestionController.java
@@ -13,7 +13,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Integer id, Model model){
+    public String question(@PathVariable("id") Long id, Model model){
         questionService.incView(id);
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("questionDTO", questionDTO);
